requirements-full.txt
# Dependencias completas para desarrollo local y scripts de update.
# Usar: pip install -r requirements-full.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
selenium>=4.0.0
webdriver-manager>=4.0.0
html5lib>=1.1
lxml>=4.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
bcchapi>=0.1.0
gunicorn>=21.2.0
reportlab>=4.0.0
pillow>=10.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
